SIM-2873: Update code to use latest fable_saga and remove custom AsyncOllama as its no longer needed.
