+++ python/mypy.ini
@@ -1,11 +1,7 @@
 [mypy]
 packages = demos, thistle_gulch
-follow_imports = skip
 ;color_output = True
 # set the stubs path
 mypy_path = ./stubs
 check_untyped_defs = True
 show_column_numbers = True
-
-[mypi-python-socketio.*]
-ignore_missing_imports = True
