+++ python/run_demos.py
@@ -2,6 +2,9 @@ import logging
 import sys
 from typing import List
 
+import demos.custom_models as custom_models
+import demos.override_action_options as action_overrides
+import demos.simulation_commands as simulation_commands
 import thistle_gulch
 import thistle_gulch.bridge as tg_bridge
 from demos import (
@@ -20,10 +23,10 @@ def main():
     # A list of available demos and the corresponding endpoint they override.
     options: List[Demo] = [
         DefaultSagaServerDemo(),
-        override_action_options.PrintActionsAndPickFirstDemo(),
-        override_action_options.SkipSagaAlwaysDoTheDefaultActionDemo(),
-        override_action_options.ReplaceContextWithYamlDumpDemo(),
-        override_action_options.UseLlama2ModelDemo(),
+        action_overrides.PrintActionsAndPickFirstDemo(),
+        action_overrides.SkipSagaAlwaysDoTheDefaultActionDemo(),
+        action_overrides.ReplaceContextWithYamlDumpDemo(),
+        custom_models.UseOllamaDemo(),
         simulation_commands.SetStartTimeDemo(),
         simulation_commands.SimulationTickDemo(),
         character_commands.EnableAgentDemo(),
