--- python/thistle_gulch/llms.py
@@ -1,138 +1,0 @@
-import json
-from typing import Any, List, Optional, AsyncIterator
-
-import aiohttp
-from langchain.callbacks.manager import (
-    CallbackManagerForLLMRun,
-    AsyncCallbackManagerForLLMRun,
-)
-from langchain.llms.ollama import Ollama
-from langchain.schema.messages import (
-    BaseMessage,
-)
-from langchain.schema.output import GenerationChunk, LLMResult
-
-
-def _stream_response_to_generation_chunk(
-    stream_response: str,
-) -> GenerationChunk:
-    """Convert a stream response to a generation chunk."""
-    parsed_response = json.loads(stream_response)
-    generation_info = parsed_response if parsed_response.get("done") is True else None
-    return GenerationChunk(
-        text=parsed_response.get("response", ""), generation_info=generation_info
-    )
-
-
-class AsyncOllama(Ollama):
-    # ...
-
-    debug_generation: bool = False
-
-    async def _create_stream(
-        self,
-        prompt: str,
-        stop: Optional[List[str]] = None,
-        **kwargs: Any,
-    ) -> AsyncIterator[str]:
-        if self.stop is not None and stop is not None:
-            raise ValueError("`stop` found in both the input and default params.")
-        elif self.stop is not None:
-            stop = self.stop
-        elif stop is None:
-            stop = []
-        params = {**self._default_params, "stop": stop, **kwargs}
-        async with aiohttp.ClientSession() as session:
-            async with session.post(
-                url=f"{self.base_url}/api/generate/",
-                headers={"Content-Type": "application/json"},
-                json={"prompt": prompt, **params},
-            ) as response:
-                if response.status != 200:
-                    optional_detail = (await response.json()).get("error")
-                    raise ValueError(
-                        f"Ollama call failed with status code {response.status}."
-                        f" Details: {optional_detail}"
-                    )
-                async for line in response.content:
-                    yield line.decode("utf-8")
-
-    async def _agenerate(
-        self,
-        prompts: List[str],
-        stop: Optional[List[str]] = None,
-        run_manager: Optional[CallbackManagerForLLMRun] = None,
-        **kwargs: Any,
-    ) -> LLMResult:
-        """Call out to Ollama's generate endpoint.
-
-        Args:
-            prompt: The prompt to pass into the model.
-            stop: Optional list of stop words to use when generating.
-
-        Returns:
-            The string generated by the model.
-
-        Example:
-            ... code-block:: python
-
-                response = ollama("Tell me a joke.")
-        """
-        # TODO: add caching here.
-        generations = []
-        for prompt in prompts:
-            final_chunk = await self._stream_with_aggregation(
-                prompt,
-                stop=stop,
-                run_manager=run_manager,
-                verbose=self.verbose,
-                **kwargs,
-            )
-            generations.append([final_chunk])
-        return LLMResult(generations=generations)
-
-    async def _astream(
-        self,
-        messages: List[BaseMessage],
-        stop: Optional[List[str]] = None,
-        run_manager: Optional[CallbackManagerForLLMRun] = None,
-        **kwargs: Any,
-    ) -> AsyncIterator[GenerationChunk]:
-        prompt = self._format_messages_as_text(messages)
-        async for stream_resp in self._create_stream(prompt, stop, **kwargs):
-            if stream_resp:
-                chunk = _stream_response_to_generation_chunk(stream_resp)
-                yield chunk
-                if run_manager:
-                    run_manager.on_llm_new_token(
-                        chunk.text,
-                        verbose=self.verbose,
-                    )
-
-    async def _stream_with_aggregation(
-        self,
-        prompt: str,
-        stop: Optional[List[str]] = None,
-        run_manager: Optional[AsyncCallbackManagerForLLMRun] = None,
-        verbose: bool = False,
-        **kwargs: Any,
-    ) -> GenerationChunk:
-        final_chunk: Optional[GenerationChunk] = None
-        async for stream_resp in self._create_stream(prompt, stop, **kwargs):
-            if stream_resp:
-                chunk = _stream_response_to_generation_chunk(stream_resp)
-                if final_chunk is None:
-                    final_chunk = chunk
-                else:
-                    if self.debug_generation:
-                        print(chunk.text, end="")
-                    final_chunk += chunk
-                if run_manager:
-                    await run_manager.on_llm_new_token(
-                        chunk.text,
-                        verbose=verbose,
-                    )
-        if final_chunk is None:
-            raise ValueError("No data received from Ollama stream.")
-
-        return final_chunk
